# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Passlinks",
    "description": "About Project",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:4200",
      "description": "Nodejs Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
    "/api/user/register": {
      "post": {
        "summary": "Registers a user.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "middleName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "aadharNumber": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created"
          },
          "500": {
            "description": "User created"
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "summary": "User login.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User login success"
          },
          "500": {
            "description": "User login failed"
          }
        }
      }
    },
    "/api/user/passportApplication": {
      "post": {
        "summary": "User applies for passport.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "DOB": {
                    "type": "string",
                    "format": "date"
                  },
                  "bloodGroup": {
                    "type": "string",
                    "enum": [
                      "O-",
                      "O+",
                      "A-",
                      "A+",
                      "B-",
                      "B+",
                      "AB-",
                      "AB+"
                    ]
                  },
                  "sex": {
                    "type": "string"
                  },
                  "nationality": {
                    "type": "string"
                  },
                  "address": {
                    "type": "object",
                    "example": {
                      "houseName": "string",
                      "streetName": "string",
                      "postOffice": "string",
                      "district": "string",
                      "state": "string",
                      "pin": "string"
                    }
                  },
                  "userImage": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Passport application submitted successfully"
          },
          "500": {
            "description": "Passport application failed"
          }
        }
      }
    },
    "/api/user/passport": {
      "get": {
        "summary": "User applies for passport.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Passport retrieval successfull",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "500": {
            "description": "Passport retrieval failed"
          }
        }
      }
    }
  }
}